refactor: changed structure and migrated to nestJS configuration
