--- nest-js/src/common/ormconfig.ts
@@ -1,14 +1,0 @@
-/* eslint-disable prettier/prettier */
-import { TypeOrmModuleOptions } from '@nestjs/typeorm';
-
-export const config: TypeOrmModuleOptions = {
-    type: 'postgres',
-    host: 'localhost',
-    port: 5433,
-    username: 'postgres',
-    password: 'korpik',
-    database: 'test',
-    synchronize: true,
-    dropSchema: false,
-    entities: ["dist/**/*.entity{.ts,.js}"],
-};
