+++ package.json
@@ -1,96 +1,76 @@
 {
-  "name": "rs-rest-service-starter",
-  "version": "1.0.0",
-  "repository": {
-    "type": "git",
-    "url": ""
-  },
-  "description": "REST service starter for RS School",
-  "main": "src/server.ts",
+  "name": "nest-js",
+  "version": "0.0.1",
+  "description": "",
+  "author": "",
+  "private": true,
+  "license": "UNLICENSED",
   "scripts": {
-    "typeorm": "node --require ts-node/register ./node_modules/typeorm/cli.js --config src/common/ormconfig.ts",
-    "create": "npm run typeorm -- migration:create -n newMigration",
-    "gener": "npm run typeorm -- migration:generate -n afterChange",
-    "mig": "ts-node --transpile-only ./node_modules/typeorm/cli.js --config src/common/ormconfig.ts migration:run",
-    "unmig": "ts-node --transpile-only ./node_modules/typeorm/cli.js --config src/common/ormconfig.ts migration:revert",
-    "doc": "jsdoc -c jsdoc.json",
-    "dev": "ts-node src/server.ts",
-    "start": "nodemon -L --exitcrash src/server.ts",
-    "lint": "eslint --ext .js,.ts src/",
-    "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
-    "test:auth": "cross-env DEBUG=rs:* TEST_MODE=auth jest --noStackTrace"
-  },
-  "keywords": [
-    "rs",
-    "school",
-    "rest",
-    "node",
-    "express",
-    "autotest",
-    "starter"
-  ],
-  "license": "ISC",
-  "engines": {
-    "node": ">=14.0.0 <15"
-  },
-  "jest": {
-    "testEnvironment": "node",
-    "setupFilesAfterEnv": [
-      "./test/setup.js"
-    ]
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "src/**/*.js": [
-      "npm run lint"
-    ]
+    "prebuild": "rimraf dist",
+    "build": "nest build",
+    "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
+    "start": "nest start",
+    "start:dev": "nest start --watch",
+    "start:debug": "nest start --debug --watch",
+    "start:prod": "node dist/main",
+    "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:cov": "jest --coverage",
+    "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
+    "test:e2e": "jest --config ./test/jest-e2e.json"
   },
   "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "cross-env": "7.0.3",
-    "dotenv": "9.0.0",
-    "express": "4.17.1",
-    "jsonwebtoken": "^8.5.1",
+    "@nestjs/common": "^7.6.15",
+    "@nestjs/core": "^7.6.15",
+    "@nestjs/mapped-types": "*",
+    "@nestjs/platform-express": "^7.6.15",
+    "@nestjs/typeorm": "^7.1.5",
+    "@types/uuid": "^8.3.1",
     "pg": "^8.6.0",
-    "postgres": "^1.0.2",
-    "stream": "^0.0.2",
-    "swagger-ui-express": "4.1.6",
-    "ts-node": "^10.0.0",
+    "reflect-metadata": "^0.1.13",
+    "rimraf": "^3.0.2",
+    "rxjs": "^6.6.6",
     "typeorm": "^0.2.34",
-    "uuid": "8.3.2",
-    "yamljs": "0.3.0"
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.2",
-    "@types/express": "^4.17.12",
-    "@types/jsonwebtoken": "^8.5.2",
-    "@types/node": "^15.6.1",
-    "@types/swagger-ui-express": "^4.1.2",
-    "@types/uuid": "^8.3.0",
-    "@types/yamljs": "^0.2.31",
-    "@typescript-eslint/eslint-plugin": "^4.25.0",
-    "@typescript-eslint/parser": "^4.25.0",
-    "chai": "4.3.4",
-    "debug": "4.3.1",
-    "dirty-chai": "2.0.1",
-    "eslint": "7.2.0",
-    "eslint-config-airbnb-base": "14.2.1",
-    "eslint-config-airbnb-typescript": "^12.3.1",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "3.4.0",
-    "husky": "4.3.8",
-    "jest": "26.6.3",
-    "jsdoc": "^3.6.7",
-    "lint-staged": "10.5.4",
-    "nodemon": "2.0.7",
-    "prettier": "2.2.1",
-    "supertest": "6.1.3",
-    "typescript": "^4.3.2"
+    "@nestjs/cli": "^7.6.0",
+    "@nestjs/schematics": "^7.3.0",
+    "@nestjs/testing": "^7.6.15",
+    "@types/express": "^4.17.11",
+    "@types/jest": "^26.0.22",
+    "@types/node": "^14.14.36",
+    "@types/supertest": "^2.0.10",
+    "@typescript-eslint/eslint-plugin": "^4.19.0",
+    "@typescript-eslint/parser": "^4.19.0",
+    "eslint": "^7.22.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-prettier": "^3.3.1",
+    "jest": "^26.6.3",
+    "prettier": "^2.2.1",
+    "supertest": "^6.1.3",
+    "ts-jest": "^26.5.4",
+    "ts-loader": "^8.0.18",
+    "ts-node": "^9.1.1",
+    "tsconfig-paths": "^3.9.0",
+    "typescript": "^4.2.3"
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "json",
+      "ts"
+    ],
+    "rootDir": "src",
+    "testRegex": ".*\\.spec\\.ts$",
+    "transform": {
+      "^.+\\.(t|j)s$": "ts-jest"
+    },
+    "collectCoverageFrom": [
+      "**/*.(t|j)s"
+    ],
+    "coverageDirectory": "../coverage",
+    "testEnvironment": "node"
   }
 }
