--- src/common/config.ts
@@ -1,20 +1,0 @@
-import dotenv from 'dotenv';
-import path from 'path';
-
-
-dotenv.config({
-  path: path.join(__dirname, '../../.env'),
-});
-
-export const config = {
-  PORT: process.env['PORT'],
-  NODE_ENV: process.env['NODE_ENV'],
-  MONGO_CONNECTION_STRING: process.env['MONGO_CONNECTION_STRING'],
-  JWT_SECRET_KEY: process.env['JWT_SECRET_KEY'],
-  AUTH_MODE: process.env['AUTH_MODE'] === 'true',
-  POSTGRES_USER: process.env['POSTGRES_USER'],
-  POSTGRES_PASSWORD: process.env['POSTGRES_PASSWORD'],
-  POSTGRES_DB: process.env['POSTGRES_DB'],
-  POSTGRES_PORT: process.env['POSTGRES_PORT'],
-  POSTGRES_HOST: process.env['POSTGRES_HOST']
-};
