--- src/resources/boards/board.column.model.ts
@@ -1,17 +1,0 @@
-import { v4 as uuid } from 'uuid';
-
-class Column {
-  id: string;
-
-  title: string;
-
-  order: number;
-
-  constructor({ id = uuid(), title = 'COLUMN', order = 0 } = {}) {
-    this.id = id;
-    this.title = title;
-    this.order = order;
-  }
-}
-
-export { Column };
