--- src/resources/login/login.service.ts
@@ -1,19 +1,0 @@
-import { getRepository } from 'typeorm';
-import bcrypt from 'bcryptjs';
-import { User } from '../../entities/user';
-
-const signUser = async (login: string, userPassword: string) => {
-    // const login = 'admin';
-    // const id = '1';
-    const foundUser = getRepository(User);
-    const admin = await foundUser.findOne({ login });
-    if (!admin) return null;
-
-    const checkBycrypt = await bcrypt.compare(userPassword, admin!.password);
-    if (checkBycrypt) {
-        return admin;
-    }
-    return null;
-};
-
-export { signUser };
