--- src/resources/tasks/_task.model.ts
@@ -1,37 +1,0 @@
-// import { v4 as uuid } from 'uuid';
-
-// class Task {
-//   id: string;
-
-//   title: string;
-
-//   order: number;
-
-//   description: string;
-
-//   userId: null;
-
-//   boardId: null;
-
-//   columnId: null;
-
-//   constructor({
-//     id = uuid(),
-//     title = 'Autotest task',
-//     order = 0,
-//     description = 'Lorem ipsum',
-//     userId = null,
-//     boardId = null,
-//     columnId = null,
-//   } = {}) {
-//     this.id = id;
-//     this.title = title;
-//     this.order = order;
-//     this.description = description;
-//     this.userId = userId;
-//     this.boardId = boardId;
-//     this.columnId = columnId;
-//   }
-// }
-
-// export { Task };
